push

+++ helper/helper_drive.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"mime/multipart"
+	"os"
 	"path/filepath"
 
 	"github.com/google/uuid"
@@ -13,15 +14,29 @@ import (
 
 var srv *drive.Service
 
+// Inisialisasi Google Drive dari ENV (untuk Heroku)
 func InitDriveService() error {
+	credsJSON := os.Getenv("GOOGLE_CREDENTIALS_JSON")
+	if credsJSON == "" {
+		return fmt.Errorf("GOOGLE_CREDENTIALS_JSON is not set")
+	}
+
+	ctx := context.Background()
+	config := []byte(credsJSON)
+
 	var err error
-	srv, err = drive.NewService(context.Background(), option.WithCredentialsFile("uploadsitemu-3488eb80fb4e.json"))
-	return err
+	srv, err = drive.NewService(ctx, option.WithCredentialsJSON(config))
+	if err != nil {
+		return fmt.Errorf("failed to initialize Google Drive: %v", err)
+	}
+
+	return nil
 }
 
+// Upload file ke Google Drive
 func UploadToDrive(file multipart.File, header *multipart.FileHeader, folderID string) (string, string, error) {
 	ext := filepath.Ext(header.Filename)
-	filename := uuidStr() + ext
+	filename := uuid.New().String() + ext
 
 	f := &drive.File{
 		Name:     filename,
@@ -31,40 +46,44 @@ func UploadToDrive(file multipart.File, header *multipart.FileHeader, folderID s
 
 	res, err := srv.Files.Create(f).Media(file).Do()
 	if err != nil {
-		return "", "", err
+		return "", "", fmt.Errorf("failed to upload to drive: %v", err)
 	}
 
 	err = makePublic(res.Id)
 	if err != nil {
-		return "", "", err
+		return "", "", fmt.Errorf("failed to make file public: %v", err)
 	}
 
 	return res.Id, getPublicURL(res.Id), nil
 }
 
+// Hapus file dari Google Drive
 func DeleteFromDrive(fileID string) error {
+	if fileID == "" {
+		return nil
+	}
 	return srv.Files.Delete(fileID).Do()
 }
 
+// Buat URL publik dari fileID
 func getPublicURL(fileID string) string {
 	return fmt.Sprintf("https://drive.google.com/uc?id=%s", fileID)
 }
 
+// Buat file publik
 func makePublic(fileID string) error {
 	perm := &drive.Permission{
-		Type: "anyone", Role: "reader",
+		Type: "anyone",
+		Role: "reader",
 	}
 	_, err := srv.Permissions.Create(fileID, perm).Do()
 	return err
 }
 
-func uuidStr() string {
-	return uuid.New().String()
-}
-
+// Ambil URL publik dari ID
 func PublicImageURLDrive(fileID string) string {
 	if fileID == "" {
 		return ""
 	}
-	return fmt.Sprintf("https://drive.google.com/uc?id=%s", fileID)
+	return getPublicURL(fileID)
 }
